add support for setting multiple hooks and initial implementation of create campaign

+++ client/app/models/CreateCampaignPayloadPartA.test.ts
@@ -5,8 +5,8 @@ describe('CreateCampaignPayloadPartA', () => {
   it('encodes and decodes a model', () => {
     const title = 'this is a brief title'
     const fundRaiseGoalInDrops = BigInt(1000000000000000000)
-    const nextMonthDateInUnixSeconds = Math.floor(
-      (Date.now() + 1000 * 60 * 60 * 24 * 30) / 1000
+    const nextMonthDateInUnixSeconds = BigInt(
+      Math.floor((Date.now() + 1000 * 60 * 60 * 24 * 30) / 1000)
     )
     const totalMilestones = 5
     const payload = new CreateCampaignPayloadPartA(
