+++ client/app/models/MilestonePayload.ts
@@ -1,14 +1,14 @@
-import { VarString, UInt32, UInt8 } from '../../util/types'
+import { VarString, UInt8, UInt64 } from '../../util/types'
 import { BaseModel, Metadata } from './BaseModel'
 
 export class MilestonePayload extends BaseModel {
   title: VarString
-  endDateInUnixSeconds: UInt32
+  endDateInUnixSeconds: UInt64
   payoutPercent: UInt8
 
   constructor(
     title: VarString,
-    endDateInUnixSeconds: UInt32,
+    endDateInUnixSeconds: UInt64,
     payoutPercent: UInt8
   ) {
     super()
@@ -20,7 +20,7 @@ export class MilestonePayload extends BaseModel {
   getMetadata(): Metadata {
     return [
       { field: 'title', type: 'varString', maxStringLength: 75 },
-      { field: 'endDateInUnixSeconds', type: 'uint32' },
+      { field: 'endDateInUnixSeconds', type: 'uint64' },
       { field: 'payoutPercent', type: 'uint8' },
     ]
   }
